+++ docs/CHANGELOG.md
@@ -5,10 +5,11 @@
 - Added a new galaxy type.
 - Added a new star type.
 - Added a new planet and moon type.
-- Added a demo limitations dialog on startup.
+- Added a toggle for immersive mode.
+- Added support for MIDI input.
 - Added a slider for drone volume.
 - Added a slider for effect volume.
-- Added support for MIDI input.
+- Added a demo limitations dialog on startup.
 - Accumulated donations when playing instruments in gallery.
 - Tracked time playing each instrument.
 - Inverted the input at The horizon.